README Update
Modified README to account for changed directory structure. Links now go
to the correct place. Abstracted functions for firebase logic now
described instead of the componentDidMount and sendMeme implementations
(which deal with mostly React specific functionality). Hopefully,
participants are more focused in on the firebase related logic in this
way.