README update
README now includes a contents section so pariticpants can easily
navigate through sections of the README. Additionally, two sections for
setup have been created (to give participants a choice in how they want
to follow along with the demo app). TODO: Decide if we want two methods
or just one (if so which one). Currently both work.

Additionally added sections for the first half of the worskhop that
explain what Firestore is and why we are using it.